+++ phase4/src/rest/jira.py
@@ -1,6 +1,7 @@
+import pytest
 import requests
 import json
-from .login_info import AuthInfo
+from .support import AuthInfo, IssueInfo
 
 
 # class for JIRA manipulations
@@ -15,7 +16,6 @@ class Jira:
         self.__search_endpoint = "api/2/search"
 
         self.__session = requests.Session()
-
 
     #
     # properties
@@ -75,3 +75,44 @@ class Jira:
     def cleanup(self):
         for key in Jira.__issues:
             self.delete_issue(key)
+
+
+# fixture to create issue before tests (for delete, update tests)
+@pytest.fixture
+def prep_issues():
+    keys = []
+    issues = [
+         IssueInfo("AQAPYTHON", "AA_issue_to_find 1", "this is a test issue for delete/update test", "Story"),
+         IssueInfo("AQAPYTHON", "AA_issue_to_find 2", "this is a test issue for delete/update test", "Bug"),
+         IssueInfo("AQAPYTHON", "AA_issue_to_find 3", "this is a test issue for delete/update test", "Bug"),
+         IssueInfo("AQAPYTHON", "AA_issue_to_find 4", "this is a test issue for delete/update test", "Bug"),
+         IssueInfo("AQAPYTHON", "AA_issue_to_find 5", "this is a test issue for delete/update test", "Bug")]
+
+    jira = Jira()
+    jira.authenticate("Alexander_Artemov", "Alexander_Artemov")
+
+    for i in issues:
+        r = jira.create_issue(i)
+        data = json.loads(r.content)
+        keys.append(data["key"])
+
+    yield
+
+    for k in keys:
+        jira.delete_issue(k)
+
+@pytest.fixture
+def prep_issue():
+    ii = IssueInfo("AQAPYTHON", "AA_issue_to_update 1", "this is a test issue for delete/update test", "Bug")
+
+    jira = Jira()
+    jira.authenticate("Alexander_Artemov", "Alexander_Artemov")
+    r = jira.create_issue(ii)
+
+    data = json.loads(r.content)
+    key = data["key"] if data["key"] else None
+
+    yield key
+
+    if key:
+        jira.delete_issue(key)
