--- phase4/src/rest/login_info.py
@@ -1,8 +1,0 @@
-class AuthInfo:
-    def __init__(self, login, password, status=None):
-        self.login = login
-        self.password = password
-        self.exp_status = status
-
-    def get_json(self):
-        return {"username": self.login, "password": self.password}
